230113 letterbox 수정 & 커밋

+++ src/main/java/com/proj/letterbox/repository/UserRepository.java
@@ -3,8 +3,8 @@ package com.proj.letterbox.repository;
 import com.proj.letterbox.model.User;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<User, Long> {
+public interface UserRepository extends CrudRepository<User, Integer> {
     public User findByEmail(String email);
-    public User findByUserCode(Long userCode);
+    public User findByUserCode(Integer userCode);
 }
